pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dockcheck"
version = "0.1.0"
description = "Agentic CI/CD runtime — safe, automated deployment with AI coding agents"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "erniesg" },
]
keywords = ["ci-cd", "agentic", "deployment", "docker", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "docker>=7.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
dockcheck = "dockcheck.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/dockcheck"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that require external services (deselect with '-m \"not slow\"')",
]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.coverage.run]
source = ["dockcheck"]

[tool.coverage.report]
show_missing = true
fail_under = 90
